Refactor: Update API service configuration
This commit updates the configuration of the API service by changing the base URL.

The previous API URL, "https://athletics-hub.uc.r.appspot.com," has been removed and replaced with a new URL, "http://35.188.51.59." This change redirects API requests to the new URL, ensuring the smooth operation of the API service.

This change addresses the need to update the base URL in order to maintain the functionality and accessibility of the API service.